Resolve some type mismatches

+++ test/main.cpp
@@ -34,7 +34,7 @@ void generate_random_file(const std::filesystem::path& file_path, uint64_t size)
         std::random_device rd;
         std::mt19937 gen(rd());
         std::uniform_int_distribution<> dis(0, 255);
-        for (uint64_t i = 0; i < buffer.size(); ++i) {
+        for (size_t i = 0; i < buffer.size(); ++i) {
             buffer[i] = static_cast<char>(dis(gen));
         }
 
@@ -167,6 +167,6 @@ int main() {
             std::cerr << "Error: " << e.what() << std::endl;
         }
     }
-    
+
     std::cerr << "Done." << std::endl;
 }
